pyproject.toml
[tool.poetry]
name = "vibe-athon2026"
version = "0.1.0"
description = "Hackathon starter project for building chatbots with FastAPI."
authors = ["Willem Ruitenbeek <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.0"
uvicorn = { extras = ["standard"], version = "^0.35.0" }
pydantic-settings = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
httpx = "^0.28.1"
ruff = "^0.12.0"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
